adding carousel for trading recom
