--- blocks/companycard/companycard.css
@@ -1,93 +1,0 @@
-.investing-ideas {
-    display: flex;
-    justify-content: space-around;
-    padding: 20px;
-    background-color: #f4f4f4;
-}
-
-.idea {
-    background-color: #ffffff;
-    border-radius: 8px;
-    padding: 20px;
-    text-align: center;
-    width: 18%; /* Adjust based on your layout preference */
-}
-
-.idea h3 {
-    color: #333333;
-    margin-bottom: 10px;
-}
-
-.profit {
-    color: green;
-    font-weight: bold;
-    margin-bottom: 5px;
-}
-
-.cmp, .target, .stop-loss {
-    color: #666666;
-    margin-bottom: 5px;
-}
-
-.btn-buy {
-    background-color: green;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-.btn-buy:hover {
-    background-color: darkgreen;
-}
-
-.view-report {
-    color: #007bff;
-    text-decoration: none;
-    display: inline-block;
-    margin-top: 10px;
-}
-
-.view-report:hover {
-    text-decoration: underline;
-}
-
-.header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 20px;
-    background-color: #003366; /* A dark shade of blue, adjust the color as needed */
-    color: white;
-}
-
-.header h1 {
-    margin: 0;
-    font-size: 24px;
-}
-
-.controls {
-    display: flex;
-}
-
-.control-btn {
-    background-color: transparent;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.control-btn.active {
-    background-color: #007bff; /* A lighter shade of blue for the active button */
-}
-
-.control-btn:not(:last-child) {
-    margin-right: 10px;
-}
-
-.control-btn:hover {
-    background-color: #0056b3; /* Darken the button on hover */
-}
