+++ styles/fonts.css
@@ -18,7 +18,7 @@
 
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 400;
   font-display: swap;
@@ -28,7 +28,7 @@
 }
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 400;
   font-display: swap;
@@ -38,7 +38,7 @@
 }
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 400;
   font-display: swap;
@@ -48,7 +48,7 @@
 }
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 500;
   font-display: swap;
@@ -58,7 +58,7 @@
 }
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 500;
   font-display: swap;
@@ -68,7 +68,7 @@
 }
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 500;
   font-display: swap;
@@ -78,7 +78,7 @@
 }
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 700;
   font-display: swap;
@@ -88,7 +88,7 @@
 }
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 700;
   font-display: swap;
@@ -98,7 +98,7 @@
 }
 
 @font-face {
-  font-family: 'Poppins';
+  font-family: Poppins;
   font-style: normal;
   font-weight: 700;
   font-display: swap;
@@ -108,7 +108,7 @@
 }
 
 @font-face {
-  font-family: "icomoon";
+  font-family: icomoon;
   src: url("../fonts/icomoon.ttf") format("truetype");
   font-display: swap;
 }
