+++ styles/styles.css
@@ -49,7 +49,7 @@
 }
 
 @font-face {
-  font-family: "icomoon";
+  font-family: icomoon;
   src: url("/fonts/icomoon.ttf") format("truetype");
 }
 
@@ -65,13 +65,15 @@
 }
 
 body {
-  font-size: var(--body-font-size-m);
   margin: 0;
-  font-family: var(--body-font-family);
-  line-height: 1.6;
-  color: var(--text-color);
-  background-color: var(--background-color);
-  display: none;
+  font: 400 14px Poppins, sans-serif;
+  color: #212529;
+  -webkit-font-smoothing: antialiased;
+  position: relative;
+  min-height: 100vh;
+  background-color: #fff;
+  text-align: left;
+  line-height: 1.5;
 }
 
 body.appear {
@@ -82,19 +84,6 @@ header {
   height: var(--nav-height);
 }
 
-
-/*h1, h2, h3,*/
-/*h4, h5, h6 {*/
-/*  font-family: var(--heading-font-family);*/
-/*  font-weight: 600;*/
-/*  line-height: 1.25;*/
-/*  margin-top: 1em;*/
-/*  margin-bottom: .5em;*/
-/*  scroll-margin: calc(var(--nav-height) + 1em);*/
-/*}*/
-
-
-
 main pre {
   background-color: var(--light-color);
   padding: 1em;
@@ -102,43 +91,6 @@ main pre {
   overflow-x: auto;
   white-space: pre;
 }
-
-/* links */
-/*a:any-link {*/
-/*  color: var(--link-color);*/
-/*  text-decoration: none;*/
-/*}*/
-
-/*a:hover {*/
-/*  text-decoration: underline;*/
-/*  color: var(--link-hover-color);*/
-/*}*/
-
-/*!* buttons *!*/
-/*a.button:any-link, button {*/
-/*  font-family: var(--body-font-family);*/
-/*  display: inline-block;*/
-/*  box-sizing: border-box;*/
-/*  text-decoration: none;*/
-/*  border: 2px solid transparent;*/
-/*  padding: 5px 30px;*/
-/*  text-align: center;*/
-/*  font-style: normal;*/
-/*  font-weight: 600;*/
-/*  cursor: pointer;*/
-/*  color: var(--background-color);*/
-/*  background-color: var(--link-color);*/
-/*  margin: 16px 0;*/
-/*  white-space: nowrap;*/
-/*  overflow: hidden;*/
-/*  text-overflow: ellipsis;*/
-/*  border-radius: 30px;*/
-/*}*/
-
-/*a.button:hover, a.button:focus, button:hover, button:focus  {*/
-/*  background-color: var(--link-hover-color);*/
-/*  cursor: pointer;*/
-/*}*/
 
 button:disabled, button:disabled:hover {
   background-color: var(--light-color);
@@ -179,11 +131,6 @@ main .section.highlight {
   background-color: var(--light-color);
 }
 
-* {
-  box-sizing: border-box;
-  outline: none;
-}
-
 div, span, strong, p, a, i {
   margin: 0;
   padding: 0;
@@ -193,7 +140,6 @@ div, span, strong, p, a, i {
 }
 
 h3 {
-  font-size: 24px;
   font-weight: 600;
   margin: 0;
   padding: 0;
@@ -203,83 +149,16 @@ h3 {
   line-height: 1.2;
 }
 
-* {
-  box-sizing: border-box;
-  outline: none;
+a {
+  color: #4b9dff;
+  text-decoration: none;
+  display: inline-block;
+  transition: 0.3s;
+  background-color: transparent;
 }
 
-.slick-dotted.slick-slider {
-  margin-bottom: 30px;
-}
-
-.slick-slider .slick-track, .slick-slider .slick-list {
-  -webkit-transform: translate3d(0, 0, 0);
-  -moz-transform: translate3d(0, 0, 0);
-  -ms-transform: translate3d(0, 0, 0);
-  -o-transform: translate3d(0, 0, 0);
-  transform: translate3d(0, 0, 0);
-}
-
-.slick-list {
-  position: relative;
-  overflow: hidden;
-  display: block;
-  margin: 0;
-  padding: 0;
-}
-
-.slick-slider {
-  position: relative;
-  display: block;
-  box-sizing: border-box;
-  -webkit-touch-callout: none;
-  -webkit-user-select: none;
-  -khtml-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-  -ms-touch-action: pan-y;
-  touch-action: pan-y;
-  -webkit-tap-highlight-color: transparent;
-}
-
-.slick-track {
-  position: relative;
-  left: 0;
-  top: 0;
-  display: flex;
-  margin-left: auto;
-  margin-right: auto;
+.align-items-center {
   align-items: center;
-  flex-wrap: nowrap;
-  transition: transform 0.5s ease;
-}
-
-.slick-track {
-  display: flex;
-  flex-wrap: nowrap; /* Ensures children are laid out in a single row */
-}
-
-.slick-track > * {
-  opacity: 0;
-}
-
-.slick-track > :first-child {
-  opacity: 1;
-}
-
-/* Tablet: first 2 children have opacity 1 and divide the width equally */
-@media (min-width: 768px) {
-  .slick-track > :nth-child(-n+2) {
-    opacity: 1;
-  }
-}
-
-/* Desktop: first 4 children have opacity 1 and equally divide the parent's width */
-@media (min-width: 1024px) {
-  .slick-track > :nth-child(-n+4) {
-    opacity: 1;
-  }
 }
 
 .row {
@@ -289,64 +168,15 @@ h3 {
   margin-left: -15px;
 }
 
-a {
-  color: #4b9dff;
-  text-decoration: none;
-  display: inline-block;
-  -webkit-transition: 0.3s;
-  -moz-transition: 0.3s;
-  -o-transition: 0.3s;
-  -ms-transition: 0.3s;
-  transition: 0.3s;
-  background-color: transparent;
-}
-
-.col-1, .col-2, .col-3, .col-4, .col-5, .col-6, .col-7, .col-8, .col-9, .col-10, .col-11, .col-12, .col, .col-auto, .col-sm-1, .col-sm-2, .col-sm-3, .col-sm-4, .col-sm-5, .col-sm-6, .col-sm-7, .col-sm-8, .col-sm-9, .col-sm-10, .col-sm-11, .col-sm-12, .col-sm, .col-sm-auto, .col-md-1, .col-md-2, .col-md-3, .col-md-4, .col-md-5, .col-md-6, .col-md-7, .col-md-8, .col-md-9, .col-md-10, .col-md-11, .col-md-12, .col-md, .col-md-auto, .col-lg-1, .col-lg-2, .col-lg-3, .col-lg-4, .col-lg-5, .col-lg-6, .col-lg-7, .col-lg-8, .col-lg-9, .col-lg-10, .col-lg-11, .col-lg-12, .col-lg, .col-lg-auto, .col-xl-1, .col-xl-2, .col-xl-3, .col-xl-4, .col-xl-5, .col-xl-6, .col-xl-7, .col-xl-8, .col-xl-9, .col-xl-10, .col-xl-11, .col-xl-12, .col-xl, .col-xl-auto {
+.col {
   position: relative;
   width: 100%;
   padding-right: 15px;
   padding-left: 15px;
 }
 
-.col-6 {
-  flex: 0 0 50%;
-  max-width: 50%;
-}
-
-.col-4 {
-  flex: 0 0 33.33333%;
-  max-width: 33.33333%;
-}
-
-.col-1, .col-2, .col-3, .col-4, .col-5, .col-6, .col-7, .col-8, .col-9, .col-10, .col-11, .col-12, .col, .col-auto, .col-sm-1, .col-sm-2, .col-sm-3, .col-sm-4, .col-sm-5, .col-sm-6, .col-sm-7, .col-sm-8, .col-sm-9, .col-sm-10, .col-sm-11, .col-sm-12, .col-sm, .col-sm-auto, .col-md-1, .col-md-2, .col-md-3, .col-md-4, .col-md-5, .col-md-6, .col-md-7, .col-md-8, .col-md-9, .col-md-10, .col-md-11, .col-md-12, .col-md, .col-md-auto, .col-lg-1, .col-lg-2, .col-lg-3, .col-lg-4, .col-lg-5, .col-lg-6, .col-lg-7, .col-lg-8, .col-lg-9, .col-lg-10, .col-lg-11, .col-lg-12, .col-lg, .col-lg-auto, .col-xl-1, .col-xl-2, .col-xl-3, .col-xl-4, .col-xl-5, .col-xl-6, .col-xl-7, .col-xl-8, .col-xl-9, .col-xl-10, .col-xl-11, .col-xl-12, .col-xl, .col-xl-auto {
-  position: relative;
-  width: 100%;
-  padding-right: 15px;
-  padding-left: 15px;
-}
-
-.col-sm-6 {
-  flex: 0 0 50%;
-  max-width: 50%;
-}
-
-@media (min-width: 576px) {
-  .col-sm-6 {
-    flex: 0 0 50%;
-    max-width: 50%;
-  }
-}
-
-@media (min-width: 768px) {
-  .col-md-6 {
-    flex: 0 0 50%;
-    max-width: 50%;
-  }
-}
-
-
-[class^="icon-"], [class*=" icon-"] {
-  font-family: "icomoon" !important;
+.icon {
+  font-family: icomoon, sans-serif;
   speak: none;
   font-style: normal;
   font-weight: 500;
@@ -359,35 +189,8 @@ a {
   vertical-align: middle;
 }
 
-img {
-  border: 0;
-  vertical-align: middle;
-  max-width: 100%;
-}
-
-
-
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, "Noto Sans", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
-  font-size: 1rem;
-  font-weight: 400;
-  line-height: 1.5;
-  color: #212529;
-  text-align: left;
-  background-color: #fff;
-}
-
-body {
-  font: 400 14px "Poppins", sans-serif;
-  color: #000000;
-  -webkit-font-smoothing: antialiased;
-  position: relative;
-  min-height: 100vh;
-}
-
 p {
-  font: 400 16px/24px "Poppins", sans-serif;
+  font: 400 16px/24px Poppins, sans-serif;
   color: #4a4a4a;
   margin-bottom: 20px;
 }
